Use a service account to get Trigger names
Previously the trigger ids we being used as the github context. It
did the job but it was pretty opaque; you could see that multiple
triggers fired, but could only see the ids not the names you assigned.

Using the cloudbuild api lets us get the actual trigger names, which we
can create meaningful context names on github PRs.