Prevent retries for malformed messages.
Malformed messages are acked with a 204 to prevent retries.
This insight is courtesy of this little bit of sanity checking:
```
gcloud pubsub topics publish cloud-builds --message "hello"
```
While that seems innocent enough, it led to 2 days of retries to the
cloud run service which didn't respond with a status code that would
stop the madness.